Float -> Double

+++ lib-esqueleto/PagilaEsqueleto.hs
@@ -101,9 +101,9 @@ $( share
             languageId Int64 sql=language_id
             originalLanguageId Int64 Maybe sql=original_language_id
             rentalDuration Int64 sql=rental_duration
-            rentalRate Float sql=rental_rate
+            rentalRate Double sql=rental_rate
             length Int64 Maybe sql=length
-            replacementCost Float sql=replacement_cost
+            replacementCost Double sql=replacement_cost
             lastUpdate UTCTime sql=last_update
             deriving Show
         FilmActor sql=film_actor
@@ -132,7 +132,7 @@ $( share
             customerId Int64 sql=customer_id
             staffId Int64 sql=staff_id
             rentalId Int64 sql=rental_id
-            amount Float sql=amount
+            amount Double sql=amount
             paymentDate UTCTime sql=payment_date
             deriving Show
         Rental sql=rental
